Add withRegistry HigherOrderComponent
